+++ astropy/modeling/functional_models.py
@@ -837,12 +837,6 @@ class Lorentz1D(Fittable1DModel):
     x_0 = Parameter(default=0)
     fwhm = Parameter(default=1)
 
-    @property
-    def stddev(self):
-        """Standard deviation based on FWHM."""
-        from ..stats.funcs import gaussian_fwhm_to_sigma
-        return self.fwhm * gaussian_fwhm_to_sigma
-
     @staticmethod
     def evaluate(x, amplitude, x_0, fwhm):
         """One dimensional Lorentzian model function"""
@@ -860,19 +854,19 @@ class Lorentz1D(Fittable1DModel):
         d_fwhm = 2 * amplitude * d_amplitude / fwhm * (1 - d_amplitude)
         return [d_amplitude, d_x_0, d_fwhm]
 
-    def bounding_box(self, factor=50):
+    def bounding_box(self, factor=25):
         """Tuple defining the default ``bounding_box`` limits,
         ``(x_low, x_high)``.
 
         Parameters
         ----------
         factor : float
-            The multiple of standard deviation used to define the limits.
+            The multiple of FWHM used to define the limits.
             Similar to `Gaussian1D`.
 
         """
         x0 = self.x_0.value
-        dx = factor * self.stddev
+        dx = factor * self.fwhm
 
         return (x0 - dx, x0 + dx)